+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## RudeHash 7.0 (TBD)
 
+* Add support for Garlicoin
+* Add support for ccminer-allium
 * Validate pool's IP address before using it
 
 ## RudeHash 6.0 (2018-03-14)